requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
django-cors-headers==3.13.0
django-filter==22.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
phonenumbers==8.13.6
phonenumberslite==8.13.6
PyJWT==1.7.1
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
